separate UsersRepository into its interface and implementation
